chore: api 파일 정리, get post, delete 삭제
