--- src/api/delete.ts
@@ -1,5 +1,0 @@
-import { DELETE } from "@/utils/axios";
-
-export const onWithDrawal = async () => {
-  await DELETE(`/user/remove`);
-};
