--- src/api/post.ts
@@ -1,5 +1,0 @@
-import { POST } from "@/utils/axios";
-
-export const onLogin = async (body: any) => {
-  await POST(`/user/login`, body);
-};
