+++ src/api/user.ts
@@ -1,8 +1,12 @@
-import { GET, POST } from "@/utils/axios";
+import { GET, POST, DELETE } from "@/utils/axios";
 
 // 유저정보 반환
 export const getUserProfile = async () => {
   return await GET("/users/profile");
+};
+
+export const onLogin = async (body: any) => {
+  await POST(`/user/login`, body);
 };
 
 // 건의사항 작성
@@ -13,4 +17,12 @@ export const userReview = async (body: any) => {
 // 내 건의사항 조회
 export const getMyReview = async () => {
   return await GET("/users/review");
+};
+
+export const onWithDrawal = async () => {
+  await DELETE(`/user/remove`);
+};
+
+export const onLogOut = async () => {
+  await GET(`/user/logout`);
 };
